\name{GraphPCA-package}
\alias{GraphPCA-package}
\alias{GraphPCA}
\docType{package}
\title{
Graphical Tools of Histogram PCA.
}
\description{
Histogram principal components analysis is the generalization of the PCA, Histogram data are adapted to design complex and Big data which histograms used as variables (Big Data adapter). Functions implemented provides numerical and graphical tools of an extension of PCA.
}
\details{
\tabular{ll}{
Package: \tab GraphPCA\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2018-04-04\cr
License: \tab GPL (>= 2)\cr
}
PrepHistogram, HistPCA, Visu.
}
\author{
Brahim Brahim <brahim.brahim@bigdatavisualizations.com> and Sun Makosso-Kallyth <makosso.sun@gmail.com>

}
\references{
Sun Makosso Kallyth (2016) principal axes analysis of symbolic histogram variables statistical snalysis and data mining.
Edwin Diday, Sun Makosso Kallyth (2012) adaptation of interval PCA to symbolic histogram variables data analysis and classification.

Billard, L. and E. Diday (2006). Symbolic Data Analysis: conceptual statistics and data Mining.
Berlin: Wiley series in computational statistics.

Diday, E., Rodriguez O. and Winberg S. (2000). Generalization of the Principal Components Analysis
to Histogram Data, 4th European Conference on Principles and Practice of Knowledge Discovery in Data Bases, September 12-16, 2000, Lyon, France.

Donoho, D.,  Ramos, E. (1982). Primdata: Data Sets for Use With PRIM-H. Version for
second (15-18, Aug, 1983) Exposition of Statistical
Graphics Technology, by American Statistical
Association.

Le-Rademacher J., Billard L. (2013). Principal component histograms from interval-valued
observations, Computational Statistics, v.28 n.5, p.2117-2138.

Makosso-Kallyth S. and Diday E. (2012). Adaptation of interval PCA to symbolic histogram 
variables, Advances in Data Analysis and Classification July, Volume 6, Issue 2, pp 147-159. 

}


\keyword{ PCA, Histogram variable, Big Data, Data Visualization, Data Analysis, }




